troca do desconto de bigDecimal pra Double, correcao nos retornos de crud da campaign pra EntityStatus no campaignController, atualizacao das funcoes de campaignService

+++ cashsplash/src/main/java/com/example/cashsplash/models/Campaign.java
@@ -20,6 +20,6 @@ public class Campaign {
     private UUID uuid;
     private String name;
     private String description;
-    private BigDecimal offValue;
+    private Double offValue;
 
 }
